+++ src/Tile.cpp
@@ -41,5 +41,5 @@ Tile::Tile(tank::Vectorf pos, int tileID, unsigned int rotation)
 
     makeGraphic <tank::Image>(*image);
     getGraphic()->setSize({64.f, 64.f});
-    getGraphic()->setClip({static_cast<int>(rotation) * 64, 0, 64 ,64});
+    getGraphic()->setClip({rotation * 64, 0, 64 ,64});
 }
